Update slider
Made the slider it's own main body item because there was so much trouble with max-width stuff because of the way I set up the layout (probably kind of problematic). Set up the div.sliderPlaceholder to create some margin at the beginning and end of the slider, which makes the snap nice and also sets the images up to extend off the screen to the right for the ux horizontal scroll hint.